+++ changelog.md
@@ -14,6 +14,7 @@ Internal improvements:
 Testing improvements:
 
  * Fix tests with ruby-head ([#674](https://github.com/arsduo/koala/pull/674))
+ * Keep supported rubies (non EOL) for CI ([#675](https://github.com/arsduo/koala/pull/675))
 
 Others: